requirements.txt
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
comtypes==1.1.11
courlan==0.6.0
dateparser==1.1.1
ffmpeg-python==0.2.0
future==0.18.2
htmldate==1.2.0
jusText==3.0.0
langcodes==3.3.0
lxml==4.8.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
pathspec==0.9.0
platformdirs==2.5.1
protobuf==3.19.4
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==303
PyYAML==6.0
regex==2022.3.2
six==1.16.0
tld==0.12.6
tomli==2.0.1
trafilatura==1.2.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
